Disable CPU enforcement if systemd-run fails
