+++ azurelinuxagent/common/event.py
@@ -72,6 +72,7 @@ class WALAEventOperation:  # pylint: disable=no-init
     ArtifactsProfileBlob = "ArtifactsProfileBlob"
     CGroupsCleanUp = "CGroupsCleanUp"
     CGroupsDebug = "CGroupsDebug"
+    CGroupsDisable = "CGroupsDisable"
     CGroupsInfo = "CGroupsInfo"
     CGroupsInitialize = "CGroupsInitialize"
     CGroupsLimitsCrossed = "CGroupsLimitsCrossed"
@@ -99,7 +100,6 @@ class WALAEventOperation:  # pylint: disable=no-init
     ImdsHeartbeat = "ImdsHeartbeat"
     Install = "Install"
     InitializeHostPlugin = "InitializeHostPlugin"
-    InvokeCommandUsingSystemd = "InvokeCommandUsingSystemd"
     Log = "Log"
     LogCollection = "LogCollection"
     OSInfo = "OSInfo"
