+++ tests/common/test_cgroupapi.py
@@ -22,15 +22,11 @@ import re
 import subprocess
 import tempfile
 
-from nose.plugins.attrib import attr
-
 from azurelinuxagent.common.cgroupapi import CGroupsApi, SystemdCgroupsApi, SYSTEMD_RUN_PATH
 from azurelinuxagent.common.cgroupstelemetry import CGroupsTelemetry
-from azurelinuxagent.common.exception import ExtensionError, ExtensionErrorCodes
-from azurelinuxagent.common.future import ustr
 from azurelinuxagent.common.utils import fileutil
 from tests.common.mock_cgroup_commands import mock_cgroup_commands
-from tests.tools import AgentTestCase, patch, skip_if_predicate_false, is_systemd_present, i_am_root, mock_sleep
+from tests.tools import AgentTestCase, patch, skip_if_predicate_false, is_systemd_present, mock_sleep
 from tests.utils.cgroups_tools import CGroupsTools
 
 
@@ -259,217 +255,6 @@ class SystemdCgroupsApiTestCase(AgentTestCase):
 
                 self.assertEqual(1, len(extension_calls), "The extension should have been invoked exactly once")
                 self.assertIn("systemd-run --unit=Microsoft.Compute.TestExtension_1.2.3", extension_calls[0], "The extension should have been invoked using systemd")
-
-    @patch('time.sleep', side_effect=lambda _: mock_sleep())
-    def test_start_extension_command_should_invoke_the_command_directly_if_systemd_fails(self, _):
-        original_popen = subprocess.Popen
-
-        def mock_popen(command, *args, **kwargs):
-            if command.startswith('systemd-run'):
-                # Inject a syntax error to the call
-                command = command.replace('systemd-run', 'systemd-run syntax_error')
-            return original_popen(command, *args, **kwargs)
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as output_file:
-            with patch("azurelinuxagent.common.cgroupapi.add_event") as mock_add_event:
-                with patch("azurelinuxagent.common.cgroupapi.subprocess.Popen", side_effect=mock_popen) as popen_patch:
-                    CGroupsTelemetry.reset()
-
-                    command = "echo TEST_OUTPUT"
-
-                    command_output = SystemdCgroupsApi().start_extension_command(
-                        extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                        command=command,
-                        timeout=300,
-                        shell=True,
-                        cwd=self.tmp_dir,
-                        env={},
-                        stdout=output_file,
-                        stderr=output_file)
-
-                    args, kwargs = mock_add_event.call_args
-                    self.assertIn("Failed to run systemd-run for unit Microsoft.Compute.TestExtension_1.2.3",
-                                  kwargs['message'])
-                    self.assertIn("Failed to find executable syntax_error: No such file or directory",
-                                  kwargs['message'])
-                    self.assertEqual(False, kwargs['is_success'])
-                    self.assertEqual('InvokeCommandUsingSystemd', kwargs['op'])
-
-                    extension_calls = [args[0] for (args, _) in popen_patch.call_args_list if command in args[0]]
-
-                    self.assertEqual(2, len(extension_calls), "The extension should have been invoked exactly twice")
-                    self.assertIn("systemd-run --unit=Microsoft.Compute.TestExtension_1.2.3", extension_calls[0],
-                                  "The first call to the extension should have used systemd")
-                    self.assertEqual(command, extension_calls[1],
-                                      "The second call to the extension should not have used systemd")
-
-                    self.assertEqual(len(CGroupsTelemetry._tracked), 0, "No cgroups should have been created")  # pylint: disable=protected-access
-
-                    self.assertIn("TEST_OUTPUT\n", command_output, "The test output was not captured")
-
-    @patch('time.sleep', side_effect=lambda _: mock_sleep())
-    def test_start_extension_command_should_invoke_the_command_directly_if_systemd_times_out(self, _):
-        # Systemd has its own internal timeout which is shorter than what we define for extension operation timeout.
-        # When systemd times out, it will write a message to stderr and exit with exit code 1.
-        # In that case, we will internally recognize the failure due to the non-zero exit code, not as a timeout.
-        original_popen = subprocess.Popen
-        systemd_timeout_command = "echo 'Failed to start transient scope unit: Connection timed out' >&2 && exit 1"
-
-        def mock_popen(*args, **kwargs):
-            # If trying to invoke systemd, mock what would happen if systemd timed out internally:
-            # write failure to stderr and exit with exit code 1.
-            new_args = args
-            if "systemd-run" in args[0]:
-                new_args = (systemd_timeout_command,)
-
-            return original_popen(new_args, **kwargs)
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stdout:
-            with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stderr:
-                with patch("azurelinuxagent.common.cgroupapi.subprocess.Popen", side_effect=mock_popen) as popen_patch:
-                    CGroupsTelemetry.reset()
-
-                    SystemdCgroupsApi().start_extension_command(
-                        extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                        command="echo 'success'",
-                        timeout=300,
-                        shell=True,
-                        cwd=self.tmp_dir,
-                        env={},
-                        stdout=stdout,
-                        stderr=stderr)
-
-                    extension_calls = [args[0] for (args, _) in popen_patch.call_args_list if "echo 'success'" in args[0]]
-
-                    self.assertEqual(2, len(extension_calls), "The extension should have been invoked exactly twice")
-                    self.assertIn("systemd-run --unit=Microsoft.Compute.TestExtension_1.2.3", extension_calls[0], "The first call to the extension should have used systemd")
-                    self.assertEqual("echo 'success'", extension_calls[1], "The second call to the extension should not have used systemd")
-
-                    self.assertEqual(len(CGroupsTelemetry._tracked), 0, "No cgroups should have been created")  # pylint: disable=protected-access
-
-    @attr('requires_sudo')
-    @patch("azurelinuxagent.common.cgroupapi.add_event")
-    @patch('time.sleep', side_effect=lambda _: mock_sleep())
-    def test_start_extension_command_should_not_use_fallback_option_if_extension_fails(self, *args):
-        self.assertTrue(i_am_root(), "Test does not run when non-root")
-        command = "ls folder_does_not_exist"
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stdout:
-            with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stderr:
-                with patch("azurelinuxagent.common.cgroupapi.subprocess.Popen", wraps=subprocess.Popen) as popen_patch:
-                    with self.assertRaises(ExtensionError) as context_manager:
-                        SystemdCgroupsApi().start_extension_command(
-                            extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                            command=command,
-                            timeout=300,
-                            shell=True,
-                            cwd=self.tmp_dir,
-                            env={},
-                            stdout=stdout,
-                            stderr=stderr)
-
-                    extension_calls = [args[0] for (args, _) in popen_patch.call_args_list if command in args[0]]
-
-                    self.assertEqual(1, len(extension_calls), "The extension should have been invoked exactly once")
-                    self.assertIn("systemd-run --unit=Microsoft.Compute.TestExtension_1.2.3", extension_calls[0],
-                                  "The first call to the extension should have used systemd")
-
-                    self.assertEqual(context_manager.exception.code, ExtensionErrorCodes.PluginUnknownFailure)
-                    self.assertIn("Non-zero exit code", ustr(context_manager.exception))
-                    # The scope name should appear in the process output since systemd-run was invoked and stderr
-                    # wasn't truncated.
-                    self.assertIn("Microsoft.Compute.TestExtension_1.2.3", ustr(context_manager.exception))
-
-    @attr('requires_sudo')
-    @patch("azurelinuxagent.common.cgroupapi.add_event")
-    @patch('time.sleep', side_effect=lambda _: mock_sleep())
-    @patch("azurelinuxagent.common.utils.extensionprocessutil.TELEMETRY_MESSAGE_MAX_LEN", 5)
-    def test_start_extension_command_should_not_use_fallback_option_if_extension_fails_with_long_output(self, *args):
-        self.assertTrue(i_am_root(), "Test does not run when non-root")
-
-        long_output = "a"*20  # large enough to ensure both stdout and stderr are truncated
-        long_stdout_stderr_command = "echo {0} && echo {0} >&2 && ls folder_does_not_exist".format(long_output)
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stdout:
-            with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stderr:
-                with patch("azurelinuxagent.common.cgroupapi.subprocess.Popen", wraps=subprocess.Popen) as popen_patch:
-                    with self.assertRaises(ExtensionError) as context_manager:
-                        SystemdCgroupsApi().start_extension_command(
-                            extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                            command=long_stdout_stderr_command,
-                            timeout=300,
-                            shell=True,
-                            cwd=self.tmp_dir,
-                            env={},
-                            stdout=stdout,
-                            stderr=stderr)
-
-                    extension_calls = [args[0] for (args, _) in popen_patch.call_args_list if long_stdout_stderr_command in args[0]]
-
-                    self.assertEqual(1, len(extension_calls), "The extension should have been invoked exactly once")
-                    self.assertIn("systemd-run --unit=Microsoft.Compute.TestExtension_1.2.3", extension_calls[0],
-                                  "The first call to the extension should have used systemd")
-
-                    self.assertEqual(context_manager.exception.code, ExtensionErrorCodes.PluginUnknownFailure)
-                    self.assertIn("Non-zero exit code", ustr(context_manager.exception))
-                    # stdout and stderr should have been truncated, so the scope name doesn't appear in stderr
-                    # even though systemd-run ran
-                    self.assertNotIn("Microsoft.Compute.TestExtension_1.2.3", ustr(context_manager.exception))
-
-    @attr('requires_sudo')
-    @patch("azurelinuxagent.common.cgroupapi.add_event")
-    def test_start_extension_command_should_not_use_fallback_option_if_extension_times_out(self, *args):  # pylint: disable=unused-argument
-        self.assertTrue(i_am_root(), "Test does not run when non-root")
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stdout:
-            with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stderr:
-                with patch("azurelinuxagent.common.utils.extensionprocessutil.wait_for_process_completion_or_timeout",
-                           return_value=[True, None]):
-                    with patch("azurelinuxagent.common.cgroupapi.SystemdCgroupsApi._is_systemd_failure",
-                               return_value=False):
-                        with self.assertRaises(ExtensionError) as context_manager:
-                            SystemdCgroupsApi().start_extension_command(
-                                extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                                command="date",
-                                timeout=300,
-                                shell=True,
-                                cwd=self.tmp_dir,
-                                env={},
-                                stdout=stdout,
-                                stderr=stderr)
-
-                        self.assertEqual(context_manager.exception.code,
-                                          ExtensionErrorCodes.PluginHandlerScriptTimedout)
-                        self.assertIn("Timeout", ustr(context_manager.exception))
-
-    @patch('time.sleep', side_effect=lambda _: mock_sleep())
-    def test_start_extension_command_should_capture_only_the_last_subprocess_output(self, _):
-        original_popen = subprocess.Popen
-
-        def mock_popen(*args, **kwargs):
-            # Inject a syntax error to the call
-            systemd_command = args[0].replace('systemd-run', 'systemd-run syntax_error')
-            new_args = (systemd_command,)
-            return original_popen(new_args, **kwargs)
-
-        expected_output = "[stdout]\n{0}\n\n\n[stderr]\n"
-
-        with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stdout:
-            with tempfile.TemporaryFile(dir=self.tmp_dir, mode="w+b") as stderr:
-                with patch("azurelinuxagent.common.cgroupapi.add_event"):
-                    with patch("azurelinuxagent.common.cgroupapi.subprocess.Popen", side_effect=mock_popen):
-                        # We expect this call to fail because of the syntax error
-                        process_output = SystemdCgroupsApi().start_extension_command(
-                            extension_name="Microsoft.Compute.TestExtension-1.2.3",
-                            command="echo 'very specific test message'",
-                            timeout=300,
-                            shell=True,
-                            cwd=self.tmp_dir,
-                            env={},
-                            stdout=stdout,
-                            stderr=stderr)
-
-                        self.assertEqual(expected_output.format("very specific test message"), process_output)
 
 
 class SystemdCgroupsApiMockedFileSystemTestCase(_MockedFileSystemTestCase):
